1、Plan 1: When there is only a consumer, the host is the local address, and the port is obtained from the port in metadata
